Masih error

+++ contracts/SimpleTransaction.sol
@@ -12,7 +12,8 @@ contract SimpleContract {
     }
 
     function getBalanceInEth(address addr) public view returns(uint) {
-        return converter.weiToEth(addr.balance);
+        return converter.weiToEth(uint(addr.balance));
+        //return uint(addr.balance) / 1000000000000000000;
     }
 
     function sendCoin(address payable receiver, uint sentCoin) public returns(bool) {
